[GridFragmentation] Start align grid items if out of fragmentainer space
When we find an with returning out-of-fragmentainer space we'll insert
a row-breakpoint, and push it into the next fragmentainer.

However - if the item is not start aligned (center or middle) we still
might run out of space. When an item runs out of space - force it to be
start aligned.

Bug: 1318608
Change-Id: I25bde1c1923443a36adec875d1d28bd0ed67f9cc