Scheduling APIs: Use abort reason in postTask
This updates postTask to use the abort reason when rejecting the
associated promise due to a signal aborting. If no abort reason is
specified during abort(), the reason defaults to an AbortError, which is
the current behavior. Given this and the newness of the API, the impact
of this change should be minimal.

Spec PR: https://github.com/WICG/scheduling-apis/pull/53

Bug: 1272822
Change-Id: Ia1dbf93e2d7e1b2e507609d0cb6f93f67deddc4e