[BlockFragmentation] Split null/max MinimalSpaceShortage representation
Previously we'd represent the "null" MinimalSpaceShortage as LayoutUnit::Max().

This causes us to be "confused" when we have a space-shortage which is
also LayoutUnit::Max(). This changes the representation to use
absl::optional<LayoutUnit> for the algorithm parts, and internally
just a LayoutUnit (similar to the current layout-result), with "null"
being kIndefiniteSize.

Bug: 1314866
Change-Id: I27be96fddcc6dba260a3b7531483b168fd75e393