[@container] Upgrade parent document if needed
If you have an iframe that's inside a container, and then call gCS
on some element inside that iframe, we need to update the style,
layout-tree *and* layout for the parent frame first. Whether or not
we should do *layout* is handled by ParentLayoutUpgrade, but it
only considered viewport-dependent media queries, not @container
dependency status.

ParentLayoutUpgrade::ShouldUpgrade, also consider whether the
Node owning the iframe needs an upgrade according to the regular
NodeLayoutUpgrade rules.

Note that the WPT currently fails due to Issue 1321010, but it's
still a useful test for this CL since it would crash without the fix.

Fixed: 1312940
Bug: 1321010
Change-Id: I6bd554618662fbb6d7e7a505d67aaeed0a4b4f03