New 'protocol-handler' PermissionDescriptor for setPermission API
The idea of this CL is to define a new mapping between blinks's
PermissionType enum and the PermissionDescriptor structure defined in by the Permission [1] interface.

The new mapping will be useful to provide permission automation for the
testing of the CustomHandler HTML API [2]. There is an ongoing
discussion in with the WPT community in the issue #26819 [3], where it
seems to be enough agreement on the idea of using the webdriver
setPermission API to grant or deny permission for the
registerProtocolHandler method, avoiding the user interaction.

This CL tries to cover all the end points of the Permission interface,
either the InternalPermission class or through the CRDTP.

[1] https://w3c.github.io/permissions/#permissions-interface
[2] https://html.spec.whatwg.org/multipage/system-state.html#custom-handlers
[3] https://github.com/web-platform-tests/wpt/issues/26819

Bug: 1321073

Change-Id: I289bc89bd1be31b18fd91ed3d81c6dd5fe853a0c