Add CSP frame-ancestor regression test.
There was already a similar test:
https://wpt.fyi/results/content-security-policy/frame-ancestors/frame-ancestors-nested-cross-in-sandboxed-cross-url-block.html
but it was using a CSP source with URLs, not 'self'.

According to https://crbug.com/1319236, Firefox behave differently and might
fails the proposed test.

I believe Chrome behavior is correct:
- 'self' is based on the document's URL
- frame-ancestors checking the ancestor's origin's URL.

As a result, a sandboed parent is never matching
CSP:frame-ancestor 'self' or <specific-url>.

See specification. Step 6.2:
https://www.w3.org/TR/CSP3/#frame-ancestors-navigation-response

I will open a bug on Mozilla bug tracker.

Bug: https://crbug.com/1319236
Change-Id: Icf95c3c7f6c0ca4412dc58d4b9481ae2da1ece91